Users have personal notes!

+++ backend/src/controllers/notes.ts
@@ -1,11 +1,15 @@
 import { RequestHandler } from "express";
-import NoteModel from "../models/Note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import NoteModel from "../models/Note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -13,9 +17,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 };
 
 export const getNote: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Not a valid object ID");
     }
@@ -24,6 +30,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     res.status(200).json(note);
@@ -43,15 +53,18 @@ export const createNote: RequestHandler<
   noteBody,
   unknown
 > = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const title = req.body.title;
   const text = req.body.text;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must have a title!");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -72,11 +85,13 @@ export const updateNote: RequestHandler<
   noteBody,
   unknown
 > = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Not a valid object ID");
     }
@@ -87,6 +102,10 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     note.title = newTitle;
@@ -101,8 +120,11 @@ export const updateNote: RequestHandler<
 };
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Not a valid object ID");
     }
@@ -111,6 +133,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     await NoteModel.findByIdAndDelete(noteId);
